requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.2
SQLAlchemy==1.4.41
Werkzeug==2.2.3
watchdog==2.1.9 
click==8.1.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
